Refactor ES
* Remove BaseESWorker, BaseESMaster

* Remove ES test function, use pycma.ff.* instead

* Move ES to root folder

* Recommand parallel ES via ProcessPoolExecutor
