--- docs/source/es.rst
@@ -1,28 +1,0 @@
-lagom.es: Evolution Strategies (ES)
-===================================
-
-.. automodule:: lagom.es
-.. currentmodule:: lagom.es
-
-.. autoclass:: BaseES
-    :members:
-    
-.. autoclass:: CMAES
-    :members:
-    
-.. autoclass:: BaseESMaster
-    :members:
-    
-.. autoclass:: BaseESWorker
-    :members:
-    
-Test functions for optimization
--------------------------------
-
-.. autofunction:: rastrigin
-
-.. autofunction:: sphere
-
-.. autofunction:: holder_table
-
-.. autofunction:: styblinski_tang
