+++ lagom/__init__.py
@@ -6,3 +6,6 @@ from .agent import BaseAgent
 from .agent import RandomAgent
 
 from .engine import BaseEngine
+
+from .es import BaseES
+from .es import CMAES
