--- lagom/es/__init__.py
@@ -1,10 +1,0 @@
-from .base_es import BaseES
-from .cma_es import CMAES
-
-from .base_es_master import BaseESMaster
-from .base_es_worker import BaseESWorker
-
-from .test_function import rastrigin
-from .test_function import sphere
-from .test_function import holder_table
-from .test_function import styblinski_tang
